got rid of __controller, fixed autocomplete bug
